requirements.txt
django==1.11.10
djangorestframework==3.6.2
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
django-extensions
Pillow
geopy
